requirements.txt
pandas>=1.5.0
openai>=1.0.0
python-dotenv>=0.19.0
openpyxl>=3.0.0 